value-pairs: move to a subdirectory
In preparation for splitting the huge value-pairs.c into smaller files,
move it to a separate directory under lib.
