+++ lib/cfg-grammar.y
@@ -341,8 +341,8 @@ extern struct _StatsOptions *last_stats_options;
 #include "rewrite/rewrite-expr.h"
 #include "rewrite/rewrite-expr-parser.h"
 #include "filter/filter-expr-parser.h"
-#include "value-pairs.h"
-#include "vptransform.h"
+#include "value-pairs/value-pairs.h"
+#include "value-pairs/vptransform.h"
 #include "file-perms.h"
 #include "block-ref-parser.h"
 #include "plugin.h"
