+++ lib/value-pairs/Makefile.am
@@ -1,0 +1,11 @@
+value_pairsincludedir			= ${pkgincludedir}/value-pairs
+
+value_pairsinclude_HEADERS = \
+	lib/value-pairs/value-pairs.h		\
+	lib/value-pairs/vptransform.h
+
+value_pairs_sources = \
+	lib/value-pairs/value-pairs.c		\
+	lib/value-pairs/vptransform.c
+
+# include lib/value-pairs/tests/Makefile.am
