+++ modules/afmongodb/afmongodb-grammar.ym
@@ -31,7 +31,7 @@
 
 #include "cfg-grammar.h"
 #include "plugin.h"
-#include "vptransform.h"
+#include "value-pairs/vptransform.h"
 
 }
