+++ modules/afmongodb/afmongodb.c
@@ -29,8 +29,7 @@
 #include "stats/stats-registry.h"
 #include "logmsg/nvtable.h"
 #include "logqueue.h"
-#include "value-pairs.h"
-#include "vptransform.h"
+#include "value-pairs/vptransform.h"
 #include "plugin.h"
 #include "plugin-types.h"
 #include "logthrdestdrv.h"
