+++ modules/afstomp/afstomp-grammar.ym
@@ -33,7 +33,7 @@
 #include "cfg-parser.h"
 #include "cfg-grammar.h"
 #include "plugin.h"
-#include "vptransform.h"
+#include "value-pairs/vptransform.h"
 
 }
