+++ modules/afstomp/afstomp.h
@@ -26,7 +26,7 @@
 #define AFSTOMP_H_INCLUDED
 
 #include "driver.h"
-#include "value-pairs.h"
+#include "value-pairs/value-pairs.h"
 
 LogDriver *afstomp_dd_new(GlobalConfig *cfg);
