+++ modules/cef/format-cef-extension.c
@@ -22,8 +22,7 @@
 #include "filter/filter-expr.h"
 #include "filter/filter-expr-parser.h"
 #include "cfg.h"
-#include "value-pairs.h"
-#include "vptransform.h"
+#include "value-pairs/value-pairs.h"
 #include "syslog-ng.h"
 #include "format-cef-extension.h"
