+++ modules/python/python-grammar.ym
@@ -26,8 +26,8 @@
 #include "python-parser.h"
 #include "python-dest.h"
 #include "python-main.h"
-#include "value-pairs.h"
-#include "vptransform.h"
+#include "value-pairs/value-pairs.h"
+#include "value-pairs/vptransform.h"
 
 }
