+++ modules/riemann/riemann-grammar.ym
@@ -24,8 +24,8 @@
 %code requires {
 
 #include "riemann-parser.h"
-#include "value-pairs.h"
-#include "vptransform.h"
+#include "value-pairs/value-pairs.h"
+#include "value-pairs/vptransform.h"
 
 }
