+++ modules/riemann/riemann.h
@@ -25,7 +25,7 @@
 #define SNG_RIEMANN_H_INCLUDED
 
 #include "driver.h"
-#include "value-pairs.h"
+#include "value-pairs/value-pairs.h"
 
 LogDriver *riemann_dd_new(GlobalConfig *cfg);
