+++ tests/unit/test_value_pairs.c
@@ -21,8 +21,8 @@
  *
  */
 
-#include "value-pairs.h"
-#include "vptransform.h"
+#include "value-pairs/value-pairs.h"
+#include "value-pairs/vptransform.h"
 #include "logmsg/logmsg.h"
 #include "apphook.h"
 #include "cfg.h"
