+++ tests/unit/test_value_pairs_walk.c
@@ -22,7 +22,7 @@
  */
 
 #include <libtest/testutils.h>
-#include <value-pairs.h>
+#include <value-pairs/value-pairs.h>
 #include <apphook.h>
 #include <plugin.h>
 #include <cfg.h>
